Separate visual and app layer in products
