+++ src/products/components/ProductListPage/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./ProductListPage";
 export * from "./ProductListPage";
+export * from "./filters";
