--- src/products/views/ProductList/messages.ts
@@ -1,33 +1,0 @@
-import { defineMessages } from "react-intl";
-
-const messages = defineMessages({
-  available: {
-    defaultMessage: "Available",
-    description: "product status"
-  },
-  hidden: {
-    defaultMessage: "Hidden",
-    description: "product is hidden"
-  },
-  outOfStock: {
-    defaultMessage: "Out Of Stock",
-    description: "product status"
-  },
-  price: {
-    defaultMessage: "Price"
-  },
-  quantity: {
-    defaultMessage: "Stock quantity",
-    description: "product"
-  },
-  visibility: {
-    defaultMessage: "Visibility",
-    description: "product visibility"
-  },
-  visible: {
-    defaultMessage: "Visible",
-    description: "product is visible"
-  }
-});
-
-export default messages;
